xip: add support for XIP packets
The eXpressive Internet Protocol (XIP) is the network layer
protocol for the eXpressive Internet Architecture (XIA), a
future Internet architecture project. The addresses in XIP are
directed acyclic graphs, so much of the code included in this
addition verifies the correctness of the DAGs and displays them
in human-readable form.

Bug: 11265
Change-Id: I948aaa73b927f8afc162d89689d184c5657f60b1